+++ backend/app/schemas/agent_runs.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from typing import Annotated, Any, Literal
 from uuid import UUID
 
-from pydantic import BaseModel, ConfigDict, Field, StringConstraints
+from pydantic import BaseModel, ConfigDict, StringConstraints
 
 # 이번 범위에서 사람이 시작할 수 있는 실행은 보고서 초안 하나다.
 AgentCode = Literal["report_writing"]
@@ -22,24 +22,6 @@ class AgentRunCreate(BaseModel):
     # 같은 키로 다시 보내면 새 실행을 만들지 않고 기존 실행을 돌려준다.
     idempotency_key: UUID
     guidance: Guidance | None = None
-
-
-class ReportDraftField(BaseModel):
-    """양식 항목 하나에 대한 제안. field_id 는 template_snapshot 의 항목 id 다."""
-
-    model_config = ConfigDict(extra="forbid")
-
-    field_id: str = Field(min_length=1, max_length=128)
-    value: str = Field(max_length=5_000)
-
-
-class ReportDraftOutput(BaseModel):
-    """LLM 이 돌려줘야 하는 구조. 검증에 실패하면 실행을 failed 로 남긴다."""
-
-    model_config = ConfigDict(extra="forbid")
-
-    fields: list[ReportDraftField] = Field(max_length=50)
-    summary: str = Field(default="", max_length=2_000)
 
 
 class AgentRunRead(BaseModel):
